+++ src/components/Signin/molecules/ButtonWrapper/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '@/components/common/atoms/Button';
+import Button from '@/components/common/Button';
 
 const ButtonWrapper = () => {
   return (
