+++ src/components/common/Button/index.stories.tsx
@@ -1,0 +1,29 @@
+import { Meta, StoryObj } from '@storybook/react';
+import Button from '.';
+
+const meta = {
+  title: 'Components/Common/Button',
+  component: Button,
+  tags: ['autodocs'],
+  argTypes: {
+    text: { control: 'text' },
+    disabled: { control: 'boolean' },
+  },
+} satisfies Meta<typeof Button>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {
+    text: 'Text',
+  },
+};
+
+export const Disabled: Story = {
+  args: {
+    text: 'Text',
+    disabled: true,
+  },
+};
