+++ src/components/common/TextField/index.stories.ts
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react';
 import TextField from './index';
 
 const config: Meta<typeof TextField> = {
-  title: 'Components/Common/atoms/TextField',
+  title: 'Components/Common/TextField',
   component: TextField,
   args: { placeholder: 'Text' },
 };
